Revert Microwatt workaround
The Icache and Dcache controllers are sending
out invalid accesses to Wishbone address space
that is not wired to peripherals, due to a old
Microwat workaround being re-added.

This commit removes the workaround and reenables
the lower address bits, re-aligning the *Cache and
Wishbone address spaces, thus allowing access to
the peripherals.

This partially resolves bug #812